# internal/http/handlers/swagger/openapi.yaml
openapi: 3.0.3

info:
  title: Beton OpenAPI
  version: 1.0.0
  summary: API documentation for the Beton backend.

servers:
  - url: http://localhost:8080
    description: Local development server.

tags:
  - name: Auth
    description: Authentication and authorization for users and staff.

  - name: Product
    description: Products and product management.

  - name: Order
    description: Client and staff orders.

  - name: Contact
    description: Contact information submitted by clients.

paths:
  /v1/products:
    get:
      operationId: getProducts
      tags:
        - Product
      description: >
        Returns products. By default, only products with displayed status
        are returned. Admins can request archived products.

      parameters:
        - name: status
          in: query
          description: Filter products by status.
          required: false
          schema:
            type: string
            enum:
              - displayed
              - archived
              - all
            default: displayed

      responses:
        "200":
          description: Products returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

        "404":
          description: Products not found.

    post:
      operationId: createProduct
      tags:
        - Product
      responses:
        "201":
          description: Product created successfully.

  /v1/orders:
    get:
      operationId: getOrders
      tags:
        - Order
      summary: Get orders
      description: Returns a paginated list of orders.

      security:
        - BearerAuth: []

      parameters:
        - name: limit
          in: query
          description: Maximum number of orders to return.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: offset
          in: query
          description: Number of orders to skip before returning results.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

      responses:
        "200":
          description: Orders returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags:
        - Order
      summary: Create order
      description: Creates a new order.

      security:
        - BearerAuth: []

      responses:
        "201":
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /v1/orders/{id}:
    delete:
      operationId: deleteOrder
      tags:
        - Order
      summary: Delete order
      description: Deletes an order by its unique identifier.

      security:
        - BearerAuth: []

      parameters:
        - name: id
          in: path
          description: Unique identifier of the order.
          required: true
          schema:
            type: integer

      responses:
        "204":
          description: Order deleted successfully.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT used to authenticate the requester.

  schemas:
    Product:
      type: object
      required:
        - id
        - price
        - title
        - type
        - status
      properties:
        id:
          type: integer
          description: Unique identifier of the product.
          example: 101

        price:
          type: integer
          description: Price in the smallest currency unit.
          example: 35000

        title:
          type: string
          description: Name of the product.
          example: Old Town Paving Stone

        type:
          type: string
          description: Product category.
          example: paving_stone

        color:
          type: string
          nullable: true
          example: Granite Gray

        status:
          $ref: "#/components/schemas/ProductStatus"

        image_urls:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"

        description:
          type: string
          nullable: true

        stock_quantity:
          type: integer
          nullable: true
          example: 1200

        weight:
          type: integer
          nullable: true
          example: 45

        rating:
          type: integer
          nullable: true
          example: 5

        size:
          $ref: "#/components/schemas/Size"

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time

    ProductStatus:
      type: string
      enum:
        - displayed
        - archived

    ProductImage:
      type: object
      required:
        - id
        - product_id
        - position
        - url
        - created_at
      properties:
        id:
          type: string
          format: uuid

        product_id:
          type: integer
          example: 101

        position:
          type: integer
          example: 1

        url:
          type: string
          format: uri
          example: https://cdn.beton.com/images/old-town-gray.jpg

        created_at:
          type: string
          format: date-time

    Size:
      type: object
      properties:
        length:
          type: integer
          example: 200

        width:
          type: integer
          example: 100

        thickness:
          type: integer
          example: 60

    Order:
      type: object
      description: Order information.
      properties:
        id:
          type: integer
          example: 1001